Move common functions from ESP32 to abstract ConnectivityManager class
